tests: replaced some tests with functional ones
Currently the tests for the cli depend to some extent on the current
implementation, both at the language level and at the function level.
This removes any dependency on internal functionality by testing at the
same level the users will use the software, that is, through a shell,
decoupling the tests from any implementation detail and testing only
about the abstraction offered to the users.

Signed-off-by: David Caro <dcaro@wikimedia.org>
Change-Id: Ie672e596e7cdf728dcd8ac5883eeaa1294e9055b